importccl: move row converter helper out of csv file
Release note: none.
