+++ pkg/ccl/importccl/read_import_proc.go
@@ -13,7 +13,12 @@ import (
 	"io"
 	"math/rand"
 	"sync"
+	"time"
 
+	"github.com/cockroachdb/cockroach/pkg/sql"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/builtins"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/transform"
+	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
 	"golang.org/x/sync/errgroup"
 
 	"github.com/cockroachdb/cockroach/pkg/ccl/storageccl"
@@ -98,47 +103,52 @@ func readInputFiles(
 			return ctx.Err()
 		default:
 		}
-		conf, err := storageccl.ExportStorageConfFromURI(dataFile)
-		if err != nil {
-			return err
-		}
-		es, err := storageccl.MakeExportStorage(ctx, conf, settings)
-		if err != nil {
-			return err
-		}
-		defer es.Close()
-		f, err := es.ReadFile(ctx, "")
-		if err != nil {
-			return err
-		}
-		defer f.Close()
-		bc := byteCounter{r: f}
-
-		wrappedProgressFn := func(finished bool) error { return nil }
-		if updateFromBytes {
-			const progressBytes = 100 << 20
-			wrappedProgressFn = func(finished bool) error {
-				// progressBytes is the number of read bytes at which to report job progress. A
-				// low value may cause excessive updates in the job table which can lead to
-				// very large rows due to MVCC saving each version.
-				if finished || bc.n > progressBytes {
-					readBytes += bc.n
-					bc.n = 0
-					if err := progressFn(float32(readBytes) / float32(totalBytes)); err != nil {
-						return err
-					}
-				}
-				return nil
-			}
-		}
-
-		if err := fileFunc(ctx, &bc, dataFileIndex, dataFile, wrappedProgressFn); err != nil {
-			return errors.Wrap(err, dataFile)
-		}
-		if updateFromFiles {
-			if err := progressFn(float32(currentFile) / float32(len(dataFiles))); err != nil {
+		if err := func() error {
+			conf, err := storageccl.ExportStorageConfFromURI(dataFile)
+			if err != nil {
 				return err
 			}
+			es, err := storageccl.MakeExportStorage(ctx, conf, settings)
+			if err != nil {
+				return err
+			}
+			defer es.Close()
+			f, err := es.ReadFile(ctx, "")
+			if err != nil {
+				return err
+			}
+			defer f.Close()
+			bc := byteCounter{r: f}
+
+			wrappedProgressFn := func(finished bool) error { return nil }
+			if updateFromBytes {
+				const progressBytes = 100 << 20
+				wrappedProgressFn = func(finished bool) error {
+					// progressBytes is the number of read bytes at which to report job progress. A
+					// low value may cause excessive updates in the job table which can lead to
+					// very large rows due to MVCC saving each version.
+					if finished || bc.n > progressBytes {
+						readBytes += bc.n
+						bc.n = 0
+						if err := progressFn(float32(readBytes) / float32(totalBytes)); err != nil {
+							return err
+						}
+					}
+					return nil
+				}
+			}
+
+			if err := fileFunc(ctx, &bc, dataFileIndex, dataFile, wrappedProgressFn); err != nil {
+				return errors.Wrap(err, dataFile)
+			}
+			if updateFromFiles {
+				if err := progressFn(float32(currentFile) / float32(len(dataFiles))); err != nil {
+					return err
+				}
+			}
+			return nil
+		}(); err != nil {
+			return err
 		}
 	}
 	return nil
@@ -153,6 +163,147 @@ func (b *byteCounter) Read(p []byte) (int, error) {
 	n, err := b.r.Read(p)
 	b.n += int64(n)
 	return n, err
+}
+
+type kvBatch []roachpb.KeyValue
+
+type rowConverter struct {
+	// stored ctx allows caching a matching Done channel between calls.
+	ctx  context.Context
+	done <-chan struct{}
+
+	// current row buf
+	datums []tree.Datum
+
+	// kv destination and current batch
+	kvCh     chan<- kvBatch
+	kvBatch  kvBatch
+	batchCap int
+
+	tableDesc *sqlbase.TableDescriptor
+
+	// The rest of these are derived from tableDesc, just cached here.
+	hidden                int
+	ri                    sqlbase.RowInserter
+	evalCtx               tree.EvalContext
+	cols                  []sqlbase.ColumnDescriptor
+	visibleCols           []sqlbase.ColumnDescriptor
+	defaultExprs          []tree.TypedExpr
+	computedIVarContainer sqlbase.RowIndexedVarContainer
+}
+
+const kvBatchSize = 1000
+
+func newRowConverter(
+	ctx context.Context, tableDesc *sqlbase.TableDescriptor, kvCh chan<- kvBatch,
+) (*rowConverter, error) {
+	c := &rowConverter{ctx: ctx, done: ctx.Done(), tableDesc: tableDesc, kvCh: kvCh}
+
+	ri, err := sqlbase.MakeRowInserter(nil /* txn */, tableDesc, nil, /* fkTables */
+		tableDesc.Columns, false /* checkFKs */, &sqlbase.DatumAlloc{})
+	if err != nil {
+		return nil, errors.Wrap(err, "make row inserter")
+	}
+	c.ri = ri
+
+	var txCtx transform.ExprTransformContext
+	evalCtx := tree.EvalContext{SessionData: &sessiondata.SessionData{Location: time.UTC}}
+	// Although we don't yet support DEFAULT expressions on visible columns,
+	// we do on hidden columns (which is only the default _rowid one). This
+	// allows those expressions to run.
+	cols, defaultExprs, err := sqlbase.ProcessDefaultColumns(tableDesc.Columns, tableDesc, &txCtx, &evalCtx)
+	if err != nil {
+		return nil, errors.Wrap(err, "process default columns")
+	}
+	c.cols = cols
+	c.defaultExprs = defaultExprs
+
+	c.visibleCols = tableDesc.VisibleColumns()
+	c.datums = make([]tree.Datum, len(c.visibleCols), len(cols))
+
+	// Check for a hidden column. This should be the unique_rowid PK if present.
+	c.hidden = -1
+	for i, col := range cols {
+		if col.Hidden {
+			if col.DefaultExpr == nil || *col.DefaultExpr != "unique_rowid()" || c.hidden != -1 {
+				return nil, errors.New("unexpected hidden column")
+			}
+			c.hidden = i
+			c.datums = append(c.datums, nil)
+		}
+	}
+	if len(c.datums) != len(cols) {
+		return nil, errors.New("unexpected hidden column")
+	}
+
+	padding := 2 * (len(tableDesc.Indexes) + len(tableDesc.Families))
+	c.batchCap = kvBatchSize + padding
+	c.kvBatch = make(kvBatch, 0, c.batchCap)
+
+	c.computedIVarContainer = sqlbase.RowIndexedVarContainer{
+		Mapping: ri.InsertColIDtoRowIndex,
+		Cols:    tableDesc.Columns,
+	}
+	return c, nil
+}
+
+func (c *rowConverter) row(fileIndex int32, rowIndex int64) error {
+	if c.hidden >= 0 {
+		// We don't want to call unique_rowid() for the hidden PK column because
+		// it is not idempotent. The sampling from the first stage will be useless
+		// during the read phase, producing a single range split with all of the
+		// data. Instead, we will call our own function that mimics that function,
+		// but more-or-less guarantees that it will not interfere with the numbers
+		// that will be produced by it. The lower 15 bits mimic the node id, but as
+		// the CSV file number. The upper 48 bits are the line number and mimic the
+		// timestamp. It would take a file with many more than 2**32 lines to even
+		// begin approaching what unique_rowid would return today, so we assume it
+		// to be safe. Since the timestamp is won't overlap, it is safe to use any
+		// number in the node id portion. The 15 bits in that portion should account
+		// for up to 32k CSV files in a single IMPORT. In the case of > 32k files,
+		// the data is xor'd so the final bits are flipped instead of set.
+		c.datums[c.hidden] = tree.NewDInt(builtins.GenerateUniqueID(fileIndex, uint64(rowIndex)))
+	}
+
+	// TODO(justin): we currently disallow computed columns in import statements.
+	var computeExprs []tree.TypedExpr
+	var computedCols []sqlbase.ColumnDescriptor
+
+	row, err := sql.GenerateInsertRow(
+		c.defaultExprs, computeExprs, c.cols, computedCols, c.evalCtx, c.tableDesc, c.datums, &c.computedIVarContainer)
+	if err != nil {
+		return errors.Wrapf(err, "generate insert row")
+	}
+	if err := c.ri.InsertRow(
+		c.ctx,
+		inserter(func(kv roachpb.KeyValue) {
+			kv.Value.InitChecksum(kv.Key)
+			c.kvBatch = append(c.kvBatch, kv)
+		}),
+		row,
+		true, /* ignoreConflicts */
+		sqlbase.SkipFKs,
+		false, /* traceKV */
+	); err != nil {
+		return errors.Wrapf(err, "insert row")
+	}
+	// If our batch is full, flush it and start a new one.
+	if len(c.kvBatch) >= kvBatchSize {
+		if err := c.sendBatch(); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (c *rowConverter) sendBatch() error {
+	select {
+	case c.kvCh <- c.kvBatch:
+	case <-c.done:
+		return c.ctx.Err()
+	}
+	c.kvBatch = make(kvBatch, 0, c.batchCap)
+	return nil
 }
 
 var csvOutputTypes = []sqlbase.ColumnType{
@@ -218,8 +369,8 @@ func (cp *readImportDataProcessor) Run(ctx context.Context, wg *sync.WaitGroup) 
 	kvCh := make(chan kvBatch)
 	sampleCh := make(chan sqlbase.EncDatumRow)
 
-	c := newCSVInputReader(ctx, cp.inputFromat.Csv, &cp.tableDesc, len(cp.tableDesc.VisibleColumns()))
-	c.start(ctx, group, kvCh)
+	c := newCSVInputReader(cp.inputFromat.Csv, &cp.tableDesc, len(cp.tableDesc.VisibleColumns()))
+	c.start(gCtx, group, kvCh)
 
 	// Read input files into kvs
 	group.Go(func() error {
